initial commit for redesigned API
