+++ Config/InstagramConfig.cs
@@ -7,9 +7,9 @@ namespace InstaSharp {
     public class InstagramConfig {
         public string APIURI { get; set; }
         public string OAuthURI { get; set; }
-        public string ClientId { get; private set; }
-        public string ClientSecret { get; private set; }
-        public string RedirectURI { get; private set; }
+        public string ClientId { get; set; }
+        public string ClientSecret { get; set; }
+        public string RedirectURI { get; set; }
 
         public InstagramConfig(string apiURI, string oauthURI, string clientId, string clientSecret, string redirectURI) {
             APIURI = apiURI;
@@ -17,6 +17,16 @@ namespace InstaSharp {
             ClientId = clientId;
             ClientSecret = clientSecret;
             RedirectURI = redirectURI;
+        }
+
+        public InstagramConfig(string apiURI)
+        {
+            APIURI = APIURI;
+        }
+
+        public InstagramConfig()
+        {
+            APIURI = "https://api.instagram.com/v1";
         }
     }
 }
